requirements.txt
PyQt5==5.15.0
Pillow==7.2.0
qrcode==6.1
six==1.15.0